Rename jla_07 to jla_06.lib
