+++ src/main/java/com/teragrep/jla_06/lib/syslog/SyslogRecordEventID.java
@@ -14,27 +14,36 @@
    See the License for the specific language governing permissions and
    limitations under the License.
  */
-package com.teragrep.jla_07.syslog;
+package com.teragrep.jla_06.lib.syslog;
 
 import com.teragrep.rlo_14.SDElement;
 import com.teragrep.rlo_14.SyslogMessage;
+import java.util.UUID;
 
-public final class SyslogRecordOrigin implements SyslogRecord {
+public final class SyslogRecordEventID implements SyslogRecord {
 
     private final SyslogRecord syslogRecord;
     private final String hostname;
 
-    public SyslogRecordOrigin(SyslogRecord syslogRecord, String hostname) {
+    public SyslogRecordEventID(SyslogRecord syslogRecord, String hostname) {
         this.syslogRecord = syslogRecord;
         this.hostname = hostname;
     }
 
     @Override
     public SyslogMessage getRecord() {
-        final SDElement origin = new SDElement("origin@48577");
-        origin.addSDParam("hostname", hostname);
+        final SDElement eventIdSDE = new SDElement("event_id@48577");
 
-        return syslogRecord.getRecord().withSDElement(origin);
+        eventIdSDE.addSDParam("hostname", hostname);
+
+        String uuid = UUID.randomUUID().toString();
+        eventIdSDE.addSDParam("uuid", uuid);
+        eventIdSDE.addSDParam("source", "source");
+
+        long unixtime = System.currentTimeMillis();
+        String epochtime = Long.toString(unixtime);
+        eventIdSDE.addSDParam("unixtime", epochtime);
+
+        return syslogRecord.getRecord().withSDElement(eventIdSDE);
     }
-
 }
