+++ src/main/java/com/teragrep/jla_06/lib/syslog/SyslogRecordSystemID.java
@@ -14,36 +14,27 @@
    See the License for the specific language governing permissions and
    limitations under the License.
  */
-package com.teragrep.jla_07.syslog;
+package com.teragrep.jla_06.lib.syslog;
 
 import com.teragrep.rlo_14.SDElement;
 import com.teragrep.rlo_14.SyslogMessage;
-import java.util.UUID;
 
-public final class SyslogRecordEventID implements SyslogRecord {
+public final class SyslogRecordSystemID implements SyslogRecord {
 
     private final SyslogRecord syslogRecord;
-    private final String hostname;
+    private final String systemID;
 
-    public SyslogRecordEventID(SyslogRecord syslogRecord, String hostname) {
+    public SyslogRecordSystemID(SyslogRecord syslogRecord, String systemID) {
         this.syslogRecord = syslogRecord;
-        this.hostname = hostname;
+        this.systemID = systemID;
     }
 
     @Override
     public SyslogMessage getRecord() {
-        final SDElement eventIdSDE = new SDElement("event_id@48577");
+        final SDElement systemIDSD = new SDElement("businessSystem@48577");
 
-        eventIdSDE.addSDParam("hostname", hostname);
+        systemIDSD.addSDParam("systemId", systemID);
 
-        String uuid = UUID.randomUUID().toString();
-        eventIdSDE.addSDParam("uuid", uuid);
-        eventIdSDE.addSDParam("source", "source");
-
-        long unixtime = System.currentTimeMillis();
-        String epochtime = Long.toString(unixtime);
-        eventIdSDE.addSDParam("unixtime", epochtime);
-
-        return syslogRecord.getRecord().withSDElement(eventIdSDE);
+        return syslogRecord.getRecord().withSDElement(systemIDSD);
     }
 }
